# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/invoices/templates": {
      "get": {
        "operationId": "InvoicesController_listTemplates",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Templates de facture",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceTemplateResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lister les templates de facture configurés",
        "tags": [
          "Factures"
        ]
      }
    },
    "/invoices/templates/{type}": {
      "get": {
        "operationId": "InvoicesController_getTemplate",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Template trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceTemplateResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer le template d'un type de facture",
        "tags": [
          "Factures"
        ]
      },
      "put": {
        "operationId": "InvoicesController_upsertTemplate",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertInvoiceTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Template upserté",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceTemplateResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer ou mettre à jour le template d'un type de facture",
        "tags": [
          "Factures"
        ]
      }
    },
    "/invoices": {
      "get": {
        "operationId": "InvoicesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Recherche textuelle",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filtre de statut",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filtre de type / origine / role selon le module",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "description": "Date debut (ISO). Champ cible selon le module (createdAt, orderDate, issueDate...).",
            "schema": {
              "example": "2026-01-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "description": "Date fin (ISO). Champ cible selon le module (createdAt, orderDate, issueDate...).",
            "schema": {
              "example": "2026-12-31T23:59:59.999Z",
              "type": "string"
            }
          },
          {
            "name": "clientId",
            "required": false,
            "in": "query",
            "description": "Filtre client (avance)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ownership",
            "required": false,
            "in": "query",
            "description": "Filtre propriété produit (COMPANY = catalogue, CLIENT = modèle client). Utilisé avec clientId = ownerClientId pour les modèles d'un client.",
            "schema": {
              "type": "string",
              "enum": [
                "COMPANY",
                "CLIENT"
              ]
            }
          },
          {
            "name": "supplierId",
            "required": false,
            "in": "query",
            "description": "Filtre fournisseur (avance)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "Filtre categorie produit (avance)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": false,
            "in": "query",
            "description": "Filtre produit (avance)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "salesOrderId",
            "required": false,
            "in": "query",
            "description": "Filtre commande de vente (avance)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "componentId",
            "required": false,
            "in": "query",
            "description": "Filtre composant / matiere (avance)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "active",
            "required": false,
            "in": "query",
            "description": "Filtre actif (ex. couleurs). true/false.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Champ de tri whitelisté côté service (ex. createdAt, name, orderDate).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sens de tri",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "includeVariants",
            "required": false,
            "in": "query",
            "description": "Inclure les variantes dans la liste (false par defaut)",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "Niveau de detail des lignes retournees. compact = payload minimum, full = payload complet.",
            "schema": {
              "default": "full",
              "type": "string",
              "enum": [
                "compact",
                "full"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste paginee des factures",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/InvoiceResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lister les factures",
        "tags": [
          "Factures"
        ]
      },
      "post": {
        "operationId": "InvoicesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Facture créée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer une facture",
        "tags": [
          "Factures"
        ]
      }
    },
    "/invoices/{id}": {
      "get": {
        "operationId": "InvoicesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Facture trouvée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer une facture",
        "tags": [
          "Factures"
        ]
      },
      "patch": {
        "operationId": "InvoicesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Facture mise à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour une facture",
        "tags": [
          "Factures"
        ]
      },
      "delete": {
        "operationId": "InvoicesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Facture supprimée"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer une facture",
        "tags": [
          "Factures"
        ]
      }
    },
    "/invoices/{id}/documents": {
      "get": {
        "operationId": "InvoicesController_listDocuments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Documents de la facture"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lister les fichiers uploadés sur une facture",
        "tags": [
          "Factures"
        ]
      },
      "post": {
        "operationId": "InvoicesController_uploadDocument",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "kind": {
                    "type": "string",
                    "enum": [
                      "SIGNED",
                      "STAMPED",
                      "SIGNED_AND_STAMPED",
                      "OTHER"
                    ]
                  },
                  "note": {
                    "type": "string"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "kind",
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document uploadé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Uploader une facture signée/cachetée (PDF ou image)",
        "tags": [
          "Factures"
        ]
      }
    },
    "/invoices/{id}/documents/{documentId}/download": {
      "get": {
        "operationId": "InvoicesController_downloadDocument",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID facture",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "description": "ID document facture",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Télécharger un document uploadé (signé/cacheté)",
        "tags": [
          "Factures"
        ]
      }
    },
    "/invoices/{id}/documents/{documentId}/replace": {
      "post": {
        "operationId": "InvoicesController_replaceDocument",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remplacer un document facture (nouvelle version)",
        "tags": [
          "Factures"
        ]
      }
    },
    "/invoices/{id}/documents/{documentId}": {
      "delete": {
        "operationId": "InvoicesController_deleteDocument",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document facture supprimé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer un document facture",
        "tags": [
          "Factures"
        ]
      }
    },
    "/invoices/{id}/mark-paid": {
      "patch": {
        "operationId": "InvoicesController_markPaid",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Facture marquée payée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Marquer une facture comme intégralement payée (raccourci)",
        "tags": [
          "Factures"
        ]
      }
    },
    "/invoices/{id}/record-payment": {
      "post": {
        "description": "Ajoute un montant encaissé. Passe automatiquement en PARTIALLY_PAID ou PAID selon si le total TTC est atteint.",
        "operationId": "InvoicesController_recordPayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordPaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Paiement enregistré",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enregistrer un paiement (acompte ou solde)",
        "tags": [
          "Factures"
        ]
      }
    },
...
  },
  "info": {
    "title": "Raffia Weave Flow API",
    "description": "Fragment OpenAPI des modules contractualisés (search/reference-lookup/audit/templates/images). Document complet: GET /{SWAGGER_PATH}-json sur l’API vivante.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "InvoiceTemplateResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Absent lorsque le template par défaut non persisté est renvoyé",
            "example": "tpl123"
          },
          "type": {
            "type": "string",
            "enum": [
              "PROFORMA",
              "DEPOSIT",
              "INTERMEDIATE",
              "FINAL",
              "CREDIT_NOTE"
            ],
            "example": "FINAL"
          },
          "name": {
            "type": "string",
            "example": "Template facture finale"
          },
          "subject": {
            "type": "object",
            "example": "Facture {{invoiceNumber}} - {{companyName}}",
            "nullable": true
          },
          "body": {
            "type": "string",
            "example": "Bonjour {{clientName}},\n\nVeuillez trouver votre facture {{invoiceNumber}}."
          },
          "footer": {
            "type": "object",
            "example": "Raffia Weave Flow",
            "nullable": true
          },
          "isDefault": {
            "type": "boolean",
            "description": "true si template synthétique (non encore persisté en base)",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-01T10:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-02T10:00:00.000Z"
          }
        },
        "required": [
          "type",
          "name",
          "body"
        ]
      },
      "UpsertInvoiceTemplateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Template facture finale premium",
            "description": "Nom lisible du template utilisé pour ce type de facture"
          },
          "subject": {
            "type": "string",
            "example": "Facture {{invoiceNumber}} - {{companyName}}",
            "description": "Sujet email suggéré lors de l'envoi de la facture"
          },
          "body": {
            "type": "string",
            "example": "Bonjour {{clientName}},\n\nVeuillez trouver votre facture {{invoiceNumber}} de {{totalTtc}} {{currency}}.",
            "description": "Corps du template. Vous pouvez y mettre vos placeholders ({{invoiceNumber}}, {{clientName}}, etc.)."
          },
          "footer": {
            "type": "string",
            "example": "Raffia Weave Flow - Merci pour votre confiance.",
            "description": "Pied de page du template"
          }
        },
        "required": [
          "name",
          "body"
        ]
      },
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "total": {
            "type": "number",
            "example": 125
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ]
      },
      "InvoiceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "inv123"
          },
          "invoiceNumber": {
            "type": "string",
            "example": "FAC-2026-0421"
          },
          "type": {
            "type": "string",
            "example": "FINAL"
          },
          "status": {
            "type": "string",
            "example": "ISSUED"
          },
          "totalTtc": {
            "type": "number",
            "example": 9792
          }
        },
        "required": [
          "id",
          "invoiceNumber",
          "type",
          "status",
          "totalTtc"
        ]
      },
...
      "CreateInvoiceDto": {
        "type": "object",
        "properties": {
          "invoiceNumber": {
            "type": "string",
            "example": "PRO/000188",
            "description": "Reference facture. Si absente, elle est generee automatiquement selon le type."
          },
          "type": {
            "type": "string",
            "example": "FINAL",
            "enum": [
              "PROFORMA",
              "DEPOSIT",
              "INTERMEDIATE",
              "FINAL",
              "CREDIT_NOTE"
            ]
          },
          "status": {
            "type": "string",
            "example": "ISSUED",
            "enum": [
              "DRAFT",
              "ISSUED",
              "SENT",
              "PAID",
              "PARTIALLY_PAID",
              "OVERDUE",
              "CANCELLED"
            ]
          },
          "clientId": {
            "type": "string",
            "example": "clx-client-id"
          },
          "salesOrderId": {
            "type": "string",
            "example": "clx-sales-order-id"
          },
          "issueDate": {
            "type": "string",
            "example": "2026-09-18T00:00:00.000Z"
          },
          "dueDate": {
            "type": "string",
            "example": "2026-10-18T00:00:00.000Z"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateInvoiceItemDto"
            }
          },
          "currency": {
            "type": "string",
            "example": "EUR"
          },
          "notes": {
            "type": "string",
            "example": "Facture finale collection hiver"
          }
        },
        "required": [
          "type",
          "clientId",
          "issueDate"
        ]
      },
...
    }
  }
}